working refactor, simulator is now running separate from synthetic dataset
