+++ examples/online/evaluate_off_policy_estimators.py
@@ -19,8 +19,7 @@ from obp.policy import LogisticEpsilonGreedy
 from obp.policy import LogisticTS
 from obp.policy import LogisticUCB
 from obp.simulator import calc_ground_truth_policy_value
-from obp.simulator import run_bandit_simulation
-
+from obp.utils import run_bandit_replay
 
 ope_estimators = [ReplayMethod()]
 
@@ -127,7 +126,7 @@ if __name__ == "__main__":
         # sample new data of synthetic logged bandit feedback
         bandit_feedback = dataset.obtain_batch_bandit_feedback(n_rounds=n_rounds)
         # simulate the evaluation policy
-        action_dist = run_bandit_simulation(
+        action_dist = run_bandit_replay(
             bandit_feedback=bandit_feedback, policy=evaluation_policy
         )
         # estimate the ground-truth policy values of the evaluation policy
