+++ obp/dataset/synthetic_multi.py
@@ -1,3 +1,5 @@
+
+
 # Copyright (c) Yuta Saito, Yusuke Narita, and ZOZO Technologies, Inc. All rights reserved.
 # Licensed under the Apache 2.0 License.
 
@@ -22,20 +24,16 @@ from .synthetic import SyntheticBanditDataset
 @dataclass
 class SyntheticMultiLoggersBanditDataset(SyntheticBanditDataset):
     """Class for synthesizing bandit data with multiple logging/behavior policies.
-
     Note
     -----
     By calling the `obtain_batch_bandit_feedback` method several times,
     we can resample logged bandit data from the same data generating distribution.
     This can be used to estimate confidence intervals of the performances of OPE estimators.
-
     If None is given as `behavior_policy_function`, the behavior policy will be generated from the true expected reward function. See the description of the `beta` argument, which controls the behavior policy.
-
     Parameters
     -----------
     n_actions: int
         Number of actions.
-
     betas: List[Union[int, float]]
         A list of inverse temperature parameters, which controls the optimality and entropy of the behavior policies.
         A large value leads to a near-deterministic behavior policy,
@@ -44,7 +42,6 @@ class SyntheticMultiLoggersBanditDataset(SyntheticBanditDataset):
         while a negative value leads to a sub-optimal behavior policy.
         Note that in the implementation, None is set as default, but a list must be set to use this class
         to generate synthetic data with multiple loggers.
-
     rhos: List[Union[int, float]]
         A list of propotions of strata sizes to generate logged bandit data with multiple loggers.
         For example, when `[0.1, 0.4, 0.5]` is given as `rhos`, there will be three strata, each stratum is generated by
@@ -53,54 +50,42 @@ class SyntheticMultiLoggersBanditDataset(SyntheticBanditDataset):
         Note that in the implementation, None is set as default, but a list must be set to use this class
         to generate synthetic data with multiple loggers.
         Moreover, `len(betas) == len(rhos)` should be met.
-
     dim_context: int, default=1
         Number of dimensions of context vectors.
-
     reward_type: str, default='binary'
         Type of reward variable, which must be either 'binary' or 'continuous'.
         When 'binary', rewards are sampled from the Bernoulli distribution.
         When 'continuous', rewards are sampled from the truncated Normal distribution with `scale=1`.
         The mean parameter of the reward distribution is determined by the `reward_function` specified by the next argument.
-
     reward_function: Callable[[np.ndarray, np.ndarray], np.ndarray]], default=None
         Function defining the expected reward for each given action-context pair,
         i.e., :math:`q: \\mathcal{X} \\times \\mathcal{A} \\rightarrow \\mathbb{R}`.
         If None, context **independent** expected rewards will be
         sampled from the uniform distribution automatically.
-
     reward_std: float, default=1.0
         Standard deviation of the reward distribution.
         A larger value leads to a noisier reward distribution.
         This argument is valid only when `reward_type="continuous"`.
-
     action_context: np.ndarray, default=None
          Vector representation of (discrete) actions.
          If None, one-hot representation will be used.
-
     n_deficient_actions: int, default=0
         Number of deficient actions having zero probability of being selected in the logged bandit data.
         If there are some deficient actions, the full/common support assumption is very likely to be violated,
         leading to some bias for IPW-type estimators. See Sachdeva et al.(2020) for details.
         `n_deficient_actions` should be an integer smaller than `n_actions - 1` so that there exists at least one action
         that have a positive probability of being selected by the behavior policy.
-
     random_state: int, default=12345
         Controls the random seed in sampling synthetic bandit data.
-
     dataset_name: str, default='synthetic_bandit_dataset'
         Name of the dataset.
-
     Examples
     ----------
-
     .. code-block:: python
-
         >>> from obp.dataset import (
             SyntheticMultiLoggersBanditDataset,
             logistic_reward_function
         )
-
         # generate synthetic contextual bandit feedback with 10 actions.
         >>> dataset = SyntheticBanditDataset(
                 n_actions=3,
@@ -145,18 +130,14 @@ class SyntheticMultiLoggersBanditDataset(SyntheticBanditDataset):
                 'pscore_avg': array([0.21089116, 0.20053648, 0.19505875, ..., 0.20616291, 0.19524095,
                         0.19032729])
             }
-
     References
     ------------
     Miroslav Dudík, Dumitru Erhan, John Langford, and Lihong Li.
     "Doubly Robust Policy Evaluation and Optimization.", 2014.
-
     Noveen Sachdeva, Yi Su, and Thorsten Joachims.
     "Off-policy Bandits with Deficient Support.", 2020.
-
     Aman Agarwal, Soumya Basu, Tobias Schnabel, Thorsten Joachims.
     "Effective Evaluation using Logged Bandit Feedback from Multiple Loggers.", 2018.
-
     Nathan Kallus, Yuta Saito, and Masatoshi Uehara.
     "Optimal Off-Policy Evaluation from Multiple Logging Policies.", 2020.
     """
@@ -244,17 +225,14 @@ class SyntheticMultiLoggersBanditDataset(SyntheticBanditDataset):
 
     def obtain_batch_bandit_feedback(self, n_rounds: int) -> BanditFeedback:
         """Obtain batch logged bandit data.
-
         Parameters
         ----------
         n_rounds: int
             Total data size (= sum of size of each stratum) of the synthetic logged bandit data.
-
         Returns
         ---------
         bandit_feedback_multi: BanditFeedback
             Synthesized logged bandit data with multiple loggers.
-
         """
         check_scalar(n_rounds, "n_rounds", int, min_val=1)
         contexts = self.random_.normal(size=(n_rounds, self.dim_context))
@@ -305,8 +283,7 @@ class SyntheticMultiLoggersBanditDataset(SyntheticBanditDataset):
             pi_b_avg += rho * softmax(beta * pi_b_logits)
 
         # sample rewards based on the context and action
-        factual_reward = self.sample_reward_given_expected_reward(expected_reward_)
-        rewards = factual_reward[np.arange(actions.shape[0]), actions]
+        rewards = self.sample_reward_given_expected_reward(expected_reward_, actions)
 
         return dict(
             n_rounds=n_rounds,
@@ -317,7 +294,6 @@ class SyntheticMultiLoggersBanditDataset(SyntheticBanditDataset):
             action=actions,
             position=None,  # position effect is not considered in synthetic data
             reward=rewards,
-            factual_reward=factual_reward,
             expected_reward=expected_reward_,
             stratum_idx=stratum_idx,
             pi_b=pi_b[:, :, np.newaxis],
