+++ tests/dataset/test_synthetic.py
@@ -6,9 +6,7 @@ import pytest
 from obp.dataset import SyntheticBanditDataset
 from obp.dataset.synthetic import (
     linear_behavior_policy,
-    ExponentialDelaySampler,
     _base_reward_function,
-    CoefficientDrifter,
 )
 from obp.dataset.synthetic import linear_reward_function
 from obp.dataset.synthetic import logistic_polynomial_reward_function
@@ -316,168 +314,6 @@ def test_synthetic_sample_reward_using_valid_inputs(context, action, description
     assert reward.shape == action.shape, "Invalid response of sample_reward"
 
 
-def test_synthetic_sample_results_in_sampled_delay_when_delay_function_is_given():
-    n_actions = 3
-    delay_function = ExponentialDelaySampler(
-        max_scale=100.0, random_state=12345
-    ).exponential_delay_function
-
-    dataset = SyntheticBanditDataset(
-        n_actions=n_actions,
-        reward_function=logistic_sparse_reward_function,
-        delay_function=delay_function,
-        random_state=12345,
-    )
-
-    actual_bandits_dataset = dataset.obtain_batch_bandit_feedback(n_rounds=5)
-
-    expected_round_delays = np.tile([266.0, 39.0, 21.0, 23.0, 84.0], (n_actions, 1)).T
-    assert (actual_bandits_dataset["round_delays"] == expected_round_delays).all()
-
-
-def test_synthetic_sample_results_in_sampled_delay_with_weighted_delays_per_arm():
-    n_actions = 3
-    delay_function = ExponentialDelaySampler(
-        max_scale=100.0, min_scale=10.0, random_state=12345
-    ).exponential_delay_function_expected_reward_weighted
-
-    dataset = SyntheticBanditDataset(
-        n_actions=n_actions,
-        reward_function=logistic_sparse_reward_function,
-        delay_function=delay_function,
-        random_state=12345,
-    )
-
-    actual_bandits_dataset = dataset.obtain_batch_bandit_feedback(n_rounds=5)
-
-    expected_round_delays = np.asarray(
-        [
-            [35.0, 38.0, 4.0],
-            [3.0, 84.0, 17.0],
-            [44.0, 106.0, 26.0],
-            [14.0, 138.0, 61.0],
-            [1.0, 12.0, 7.0],
-        ]
-    )
-    assert (actual_bandits_dataset["round_delays"] == expected_round_delays).all()
-
-
-def test_synthetic_sample_results_with_exponential_delay_function_has_different_delays_each_batch():
-    n_actions = 3
-
-    delay_function = ExponentialDelaySampler(
-        max_scale=1000.0, random_state=12345
-    ).exponential_delay_function
-
-    dataset = SyntheticBanditDataset(
-        n_actions=n_actions,
-        reward_function=logistic_sparse_reward_function,
-        delay_function=delay_function,
-        random_state=12345,
-    )
-
-    actual_delays_1 = dataset.obtain_batch_bandit_feedback(n_rounds=5)["round_delays"]
-    actual_delays_2 = dataset.obtain_batch_bandit_feedback(n_rounds=5)["round_delays"]
-
-    expected_round_delays_1 = np.tile(
-        [2654.0, 381.0, 204.0, 229.0, 839.0], (n_actions, 1)
-    ).T
-    expected_round_delays_2 = np.tile(
-        [906.0, 3339.0, 1059.0, 1382.0, 1061.0], (n_actions, 1)
-    ).T
-
-    assert (actual_delays_1 == expected_round_delays_1).all()
-    assert (actual_delays_2 == expected_round_delays_2).all()
-
-
-def test_synthetic_sample_results_with_exponential_delay_function_has_same_delays_each_dataset():
-    n_actions = 3
-
-    delay_function = ExponentialDelaySampler(
-        max_scale=1000.0, random_state=12345
-    ).exponential_delay_function
-
-    dataset = SyntheticBanditDataset(
-        n_actions=n_actions,
-        reward_function=logistic_sparse_reward_function,
-        delay_function=delay_function,
-        random_state=12345,
-    )
-
-    actual_delays_1 = dataset.obtain_batch_bandit_feedback(n_rounds=5)["round_delays"]
-
-    delay_function = ExponentialDelaySampler(
-        max_scale=1000.0, random_state=12345
-    ).exponential_delay_function
-
-    dataset = SyntheticBanditDataset(
-        n_actions=n_actions,
-        reward_function=logistic_sparse_reward_function,
-        delay_function=delay_function,
-        random_state=12345,
-    )
-    actual_delays_2 = dataset.obtain_batch_bandit_feedback(n_rounds=5)["round_delays"]
-
-    expected_round_delays_1 = np.tile(
-        [2654.0, 381.0, 204.0, 229.0, 839.0], (n_actions, 1)
-    ).T
-    expected_round_delays_2 = np.tile(
-        [2654.0, 381.0, 204.0, 229.0, 839.0], (n_actions, 1)
-    ).T
-
-    assert (actual_delays_1 == expected_round_delays_1).all()
-    assert (actual_delays_2 == expected_round_delays_2).all()
-
-
-def test_synthetic_sample_results_do_not_contain_reward_delay_when_delay_function_is_none():
-    dataset = SyntheticBanditDataset(
-        n_actions=3,
-        reward_function=logistic_sparse_reward_function,
-        random_state=12345,
-    )
-
-    actual_bandits_dataset = dataset.obtain_batch_bandit_feedback(n_rounds=5)
-
-    assert actual_bandits_dataset["round_delays"] is None
-
-
-def test_synthetic_sample_results_reward_delay_is_configurable_through_delay_function():
-    n_actions = 3
-
-    def trivial_delay_func(*args, **kwargs):
-        return np.asarray([1, 2, 3, 4, 5])
-
-    dataset = SyntheticBanditDataset(
-        n_actions=3,
-        reward_function=logistic_sparse_reward_function,
-        delay_function=trivial_delay_func,
-        random_state=12345,
-    )
-
-    actual_bandits_dataset = dataset.obtain_batch_bandit_feedback(n_rounds=5)
-
-    assert (actual_bandits_dataset["round_delays"] == [1, 2, 3, 4, 5]).all()
-
-
-@pytest.mark.parametrize(
-    "size, actions, random_state, expected_delays",
-    [
-        (5, 3, 12345, np.tile([266.0, 39.0, 21.0, 23.0, 84.0], (3, 1)).T),
-        (3, 3, 12345, np.tile([266.0, 39.0, 21.0], (3, 1)).T),
-        (5, 3, 54321, np.tile([243.0, 98.0, 157.0, 57.0, 79.0], (3, 1)).T),
-    ],
-)
-def test_exponential_delay_function_results_in_expected_seeded_discrete_delays(
-    size, actions, random_state, expected_delays
-):
-    delay_function = ExponentialDelaySampler(
-        max_scale=100.0, random_state=random_state
-    ).exponential_delay_function
-
-    actual_delays = delay_function(size, n_actions=actions)
-    assert (actual_delays == expected_delays).all()
-
-
 def test_synthetic_obtain_batch_bandit_feedback():
     # n_rounds
     with pytest.raises(ValueError):
@@ -533,36 +369,6 @@ def test_synthetic_obtain_batch_bandit_feedback():
             bandit_feedback["pscore"].ndim == 1
             and len(bandit_feedback["pscore"]) == n_rounds
         )
-
-
-@pytest.mark.parametrize(
-    "size, actions, expected_reward, random_state, expected_delays",
-    [
-        (
-            2,
-            2,
-            np.asarray([[1, 0.01], [0.5, 0.5]]),
-            12344,
-            np.asarray([[2.0, 55.0], [3.0, 27.0]]),
-        ),
-        (
-            2,
-            2,
-            np.asarray([[0.1, 0.2], [0.3, 0.4]]),
-            12345,
-            np.asarray([[242.0, 32.0], [15.0, 15.0]]),
-        ),
-    ],
-)
-def test_exponential_delay_function_conditioned_on_expected_reward_results_in_expected_seeded_discrete_delays(
-    size, actions, expected_reward, random_state, expected_delays
-):
-    delay_function = ExponentialDelaySampler(
-        max_scale=100.0, min_scale=10.0, random_state=random_state
-    ).exponential_delay_function_expected_reward_weighted
-
-    actual_delays = delay_function(expected_rewards=expected_reward)
-    assert (actual_delays == expected_delays).all()
 
 
 # expected_reward, action_dist, description
@@ -786,430 +592,3 @@ def test_base_reward_create_a_matrix_with_expected_rewards_with_identical_expect
     )
 
     assert np.allclose(actual_expected_rewards, expected_expected_rewards)
-
-
-def test_coefficient_tracker_can_shift_expected_rewards_with_syntethic_dataset_generator():
-    drifter = CoefficientDrifter(drift_interval=3)
-
-    dataset = SyntheticBanditDataset(
-        n_actions=3,
-        dim_context=4,
-        reward_type="binary",  # "binary" or "continuous"
-        reward_function=logistic_sparse_reward_function,
-        coef_function=drifter.get_coefficients,
-        random_state=12345,
-    )
-
-    bandit_dataset = dataset.obtain_batch_bandit_feedback(n_rounds=4)
-
-
-class MockCoefSample:
-    n_samples = 0
-
-    def fake_sample(
-        self, effective_dim_action_context, effective_dim_context, random_, **kwargs
-    ):
-        self.n_samples += 1
-        context_coef_ = self.n_samples * np.ones(effective_dim_context)
-        action_coef_ = self.n_samples * np.ones(effective_dim_action_context)
-        context_action_coef_ = self.n_samples * np.ones(
-            (effective_dim_context, effective_dim_action_context)
-        )
-
-        return context_coef_, action_coef_, context_action_coef_
-
-
-def test_coefficient_tracker_can_shift_expected_rewards_instantly_based_on_configured_intervals():
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(drift_interval=3)
-
-        context = np.asarray(
-            [
-                [1, 2],
-                [3, 2],
-                [3, 2],
-                [3, 2],
-            ]
-        )
-        action_context = np.asarray(
-            [
-                [1, 0, 0],
-                [0, 0, 1],
-                [0, 1, 0],
-            ]
-        )
-        actual_expected_rewards = _base_reward_function(
-            context,
-            action_context,
-            degree=5,
-            effective_dim_ratio=1.0,
-            coef_function=drifter.get_coefficients,
-            random_state=12345,
-        )
-
-    expected_expected_rewards = np.asarray(
-        [
-            [-7.88993004, -7.88993004, -7.88993004],  # This round has a different context and should have diff E[r]
-            [ 0.9467916 ,  0.9467916 ,  0.9467916 ],  # The next two rounds have the same context and should have identical
-            [ 0.9467916 ,  0.9467916 ,  0.9467916 ],  # E[r]
-            [ 5.99634683,  5.99634683,  5.99634683],  # This round has the same context but has experienced drift.
-        ]
-    )
-
-    assert np.allclose(actual_expected_rewards, expected_expected_rewards)
-
-
-def test_coefficient_tracker_can_shift_coefficient_instantly_based_on_configured_interval():
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        effective_dim_context = 4
-        effective_dim_action_context = 3
-        drifter = CoefficientDrifter(
-            drift_interval=3,
-            effective_dim_context=effective_dim_context,
-            effective_dim_action_context=effective_dim_action_context,
-        )
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=4)
-
-    expected_context_coef = np.asarray(
-        [
-            [2, 2, 2, 2],
-            [2, 2, 2, 2],
-            [2, 2, 2, 2],  # AFTER THIS ROUND, THE COEFS CHANGE ABRUPTLY
-            [3, 3, 3, 3],
-        ]
-    )
-
-    assert np.allclose(actual_context_coef, expected_context_coef)
-
-
-def test_coefficient_tracker_can_shift_linearly_instantly_based_on_configured_transition_period():
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=4,
-            transition_period=2,
-            effective_dim_context=2,
-            effective_dim_action_context=2,
-        )
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=8)
-
-    expected_context_coef = np.asarray(
-        [
-            [2.0, 2.0],
-            [2.0, 2.0],  # First two rounds are the same
-            [2.33333333, 2.33333333],  # Next two rounds slowly transition
-            [2.66666667, 2.66666667],
-            [3.0, 3.0],  # Now we start in the new coef again
-            [3.0, 3.0],
-            [3.33333333, 3.33333333],  # Now we start transitioning again.
-            [3.66666667, 3.66666667],
-        ]
-    )
-
-    assert np.allclose(actual_context_coef, expected_context_coef)
-
-
-def test_coefficient_tracker_can_shift_weighted_sampled_based_on_configured_transition_period():
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=4,
-            transition_period=2,
-            transition_type="weighted_sampled",
-            effective_dim_context=2,
-            effective_dim_action_context=2,
-        )
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=8)
-
-    expected_context_coef = np.asarray(
-        [
-            [2.0, 2.0],
-            [2.0, 2.0],  # First two rounds are the same
-            [2.0, 2.0],  # Next two rounds are weighted sampled
-            [3.0, 3.0],
-            [3.0, 3.0],  # Now we start in the new coef again
-            [3.0, 3.0],
-            [4.0, 4.0],  # Next two rounds are weighted sampled again
-            [4.0, 4.0],
-        ]
-    )
-
-    assert np.allclose(actual_context_coef, expected_context_coef)
-
-
-def test_coefficient_tracker_can_shift_instantly_back_and_forth_between_seasons_using_seasonality_flag():
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=2,
-            transition_period=0,
-            seasonal=True,
-            effective_dim_context=2,
-            effective_dim_action_context=2,
-        )
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=8)
-
-    expected_context_coef = np.asarray(
-        [
-            [2.0, 2.0],
-            [2.0, 2.0],
-            [3.0, 3.0],
-            [3.0, 3.0],
-            [2.0, 2.0],
-            [2.0, 2.0],
-            [3.0, 3.0],
-            [3.0, 3.0],
-        ]
-    )
-
-    assert np.allclose(actual_context_coef, expected_context_coef)
-
-
-def test_coefficient_tracker_can_shift_instantly_under_base_coeficient():
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=2,
-            transition_period=0,
-            seasonal=True,
-            base_coefficient_weight=0.8,
-            effective_dim_context=2,
-            effective_dim_action_context=2,
-        )
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=8)
-
-        expected_context_coef = np.asarray(
-            [
-                [1.2, 1.2],
-                [1.2, 1.2],
-                [1.4, 1.4],
-                [1.4, 1.4],
-                [1.2, 1.2],
-                [1.2, 1.2],
-                [1.4, 1.4],
-                [1.4, 1.4],
-            ]
-        )
-
-    assert np.allclose(actual_context_coef, expected_context_coef)
-
-
-def test_coefficient_tracker_update_coef_makes_next_coef_current_coef_and_samples_new_next_coef():
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=4, effective_dim_context=1, effective_dim_action_context=1
-        )
-
-        drifter.context_coefs[0] = [1]
-        drifter.context_coefs[1] = [2]
-
-        drifter.update_coef()
-
-        assert drifter.context_coefs[0] == [2]
-        assert np.allclose(drifter.context_coefs[1], [4.0])
-
-
-def test_coefficient_tracker_update_coef_samples_both_new_curr_and_next_on_first_pull():
-    drifter = CoefficientDrifter(
-        drift_interval=4,
-    )
-
-    assert len(drifter.context_coefs) == 0
-
-    drifter.effective_dim_context = 1
-    drifter.effective_dim_action_context = 1
-
-    drifter.update_coef()
-
-    assert drifter.context_coefs[0] is not None
-    assert drifter.context_coefs[1] is not None
-    assert not np.allclose(drifter.context_coefs[0], drifter.context_coefs[1])
-
-
-def test_coefficient_tracker_can_set_effective_dim_context_on_first_sample():
-    effective_dim_context = 4
-    effective_dim_action_context = 3
-    drifter = CoefficientDrifter(drift_interval=3)
-
-    assert drifter.effective_dim_context == None
-    assert drifter.effective_dim_action_context == None
-
-    actual_context_coef, _, _ = drifter.get_coefficients(
-        n_rounds=4,
-        effective_dim_context=effective_dim_context,
-        effective_dim_action_context=effective_dim_action_context,
-    )
-
-    assert drifter.effective_dim_context == 4
-    assert drifter.effective_dim_action_context == 3
-
-
-def test_coefficient_tracker_raises_when_effective_dimensions_are_being_changed():
-    effective_dim_context = 4
-    effective_dim_action_context = 3
-    drifter = CoefficientDrifter(
-        drift_interval=3,
-        effective_dim_context=effective_dim_context,
-        effective_dim_action_context=effective_dim_action_context,
-    )
-
-    assert drifter.effective_dim_context == 4
-    assert drifter.effective_dim_action_context == 3
-
-    with pytest.raises(
-        RuntimeError, match=r"Trying to change the effective dimensions"
-    ):
-        actual_context_coef, _, _ = drifter.get_coefficients(
-            n_rounds=4, effective_dim_context=5, effective_dim_action_context=6
-        )
-
-
-def test_coefficient_tracker_can_shift_coefficient_multiple_times_instantly_based_on_configured_interval():
-    effective_dim_context = 4
-    effective_dim_action_context = 3
-
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=2,
-            effective_dim_context=effective_dim_context,
-            effective_dim_action_context=effective_dim_action_context,
-        )
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=5)
-
-    expected_context_coef = np.asarray(
-        [
-            [2.0, 2.0, 2.0, 2.0],
-            [2.0, 2.0, 2.0, 2.0],  # AFTER THIS ROUND, THE COEFS SHOULD CHANGE
-            [3.0, 3.0, 3.0, 3.0],
-            [3.0, 3.0, 3.0, 3.0],  # AFTER THIS ROUND, THE COEFS SHOULD CHANGE AGAIN
-            [4.0, 4.0, 4.0, 4.0],
-        ]
-    )
-
-    assert np.allclose(actual_context_coef, expected_context_coef)
-
-
-def test_coefficient_tracker_keeps_track_of_shifted_coefficient_based_on_configured_interval_between_batches():
-    effective_dim_context = 4
-    effective_dim_action_context = 3
-
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=2,
-            effective_dim_context=effective_dim_context,
-            effective_dim_action_context=effective_dim_action_context,
-        )
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=3)
-
-        expected_context_coef = np.asarray(
-            [
-                [2.0, 2.0, 2.0, 2.0],
-                [2.0, 2.0, 2.0, 2.0],  # AFTER THIS ROUND, THE COEFS SHOULD CHANGE
-                [3.0, 3.0, 3.0, 3.0],
-            ]
-        )
-
-        assert np.allclose(actual_context_coef, expected_context_coef)
-
-        actual_context_coef, _, _ = drifter.get_coefficients(n_rounds=3)
-
-        expected_context_coef_2 = np.asarray(
-            [
-                [3.0, 3.0, 3.0, 3.0],  # THIS ROUND SHOULD BE THE SAME AS THE LAST ONE
-                [4.0, 4.0, 4.0, 4.0],  # HERE THE COEF SHOULD CHANGE AGAIN
-                [4.0, 4.0, 4.0, 4.0],
-            ]
-        )
-
-    assert np.allclose(actual_context_coef, expected_context_coef_2)
-
-
-def test_coefficients_can_drift_for_the_action_coefs():
-    effective_dim_context = 4
-    effective_dim_action_context = 3
-
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=2,
-            effective_dim_context=effective_dim_context,
-            effective_dim_action_context=effective_dim_action_context,
-        )
-
-        _, actual_action_coef, _ = drifter.get_coefficients(n_rounds=3)
-
-        expected_action_coef = np.asarray(
-            [
-                [2.0, 2.0, 2.0],
-                [2.0, 2.0, 2.0],  # AFTER THIS ROUND, THE COEFS SHOULD CHANGE
-                [3.0, 3.0, 3.0],
-            ]
-        )
-
-        assert np.allclose(actual_action_coef, expected_action_coef)
-
-
-
-def test_coefficients_can_drift_for_the_action_coefs():
-    effective_dim_context = 4
-    effective_dim_action_context = 3
-
-    with mock.patch(
-        "obp.dataset.synthetic.sample_random_uniform_coefficients",
-        MockCoefSample().fake_sample,
-    ):
-        drifter = CoefficientDrifter(
-            drift_interval=2,
-            effective_dim_context=effective_dim_context,
-            effective_dim_action_context=effective_dim_action_context,
-        )
-
-        _, _, actual_context_action_coef = drifter.get_coefficients(n_rounds=3)
-
-        expected_context_action_coef = np.asarray([[[2., 2., 2.],
-        [2., 2., 2.],
-        [2., 2., 2.],
-        [2., 2., 2.]],
-       [[2., 2., 2.],
-        [2., 2., 2.],
-        [2., 2., 2.],
-        [2., 2., 2.]],
-       [[3., 3., 3.],
-        [3., 3., 3.],
-        [3., 3., 3.],
-        [3., 3., 3.]]])
-
-        assert np.allclose(actual_context_action_coef, expected_context_action_coef)
-
